added jobs support

+++ MertricAgentServices/BaseService.cs
@@ -1,13 +1,10 @@
-﻿using AutoMapper;
-using DB.Interfaces;
+﻿using DB.Interfaces;
 using DB.Models;
 using MertricAgentServices.Dto;
 using MertricAgentServices.Mapper;
-using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace MertricAgentServices
@@ -23,10 +20,10 @@ namespace MertricAgentServices
         }
         protected IMetricMapper Mapper { get; }
         protected IDbRepository<TEntity> Repository { get; }
-        public virtual async Task UpdateAsync(TDto dto)
+        public virtual async Task UpdateAsync(long id, TDto dto)
         {
             _ = dto ?? throw new ArgumentException(nameof(dto));
-            var entity = Repository.Get().Where(x=>x.Id==dto.Id).FirstOrDefault();
+            var entity = Repository.Get().FirstOrDefault(s=>s.Id.Equals(id));
             Mapper.Map(dto, entity);
             await Repository.UpdateAsync(entity);
         }
@@ -36,10 +33,13 @@ namespace MertricAgentServices
             var entity = Mapper.Map<TEntity>(dto);
             await Repository.AddAsync(entity);
         }
-        public virtual async Task DeleteAsync(TDto dto)
+        public virtual async Task DeleteAsync(long id)
         {
-            _ = dto ?? throw new ArgumentException(nameof(dto));
-            var entity = Mapper.Map<TEntity>(dto);
+            var entity = Repository.Get().FirstOrDefault(s => s.Id.Equals(id));
+            if(entity == null)
+            {
+                throw new Exception();
+            }
             await Repository.DeleteAsync(entity);
         }
         public virtual List<TEntity> Get()
